[lldb/test] Fix Swift regex tests on remote Darwin platforms
The Makefile hardcoded -target $(ARCH)-apple-macosx13.0 regardless of
the actual build platform. On remote Darwin runs the test runner already
supplies the correct triple via TARGET_SWIFTFLAGS, so the override was
causing the binary to be built for macOS even when targeting iOS, leading
to a confusing "no such directory" launch failure on the device.

The test is split into one file per test method so lit can schedule them
in parallel, avoiding the sequential 20+ variant × ~20s/variant bottleneck
that caused the full suite to time out. Each test is restricted to dsym
variants with @skipIf(debug_info=no_match("dsym")) since the regex
expression evaluation is independent of debug info format and the dwarf
variants were hitting 20-second expression evaluator timeouts.

(cherry picked from commit f2bb8d022014168ca82d75232470d5de09bf7d93)
