+++ lldb/test/API/lang/swift/regex/Makefile
@@ -1,8 +1,8 @@
-ifeq "$(OS)" "Darwin"
-	OS_SWIFTFLAGS := -target $(ARCH)-apple-macosx13.0
-endif
-
 SWIFT_SOURCES := main.swift
-SWIFTFLAGS_EXTRAS := -enable-bare-slash-regex $(OS_SWIFTFLAGS)
+# -enable-bare-slash-regex enables the /pattern/ literal syntax.
+# Do not hardcode a macOS deployment target here — the test runner
+# supplies the correct target triple via TARGET_SWIFTFLAGS for both
+# local macOS and remote Darwin builds.
+SWIFTFLAGS_EXTRAS := -enable-bare-slash-regex
 
 include Makefile.rules
